requirements.txt
flask
elevenlabs
google-genai
python-dotenv
requests
moviepy
